+++ components/forms/QualityCheckAtac.vue
@@ -12,7 +12,7 @@
                 <v-card
                   v-if="qcResult !== null"
                   outlined
-                  color="blue-grey lighten-5"
+                  color="grey lighten-3"
                 >
                   <v-card-text>
                     <p class="subtitle-1 font-weight-bold text-center">
@@ -88,7 +88,7 @@
                   </v-row>
                 </v-card>
 
-                <v-card class="mt-6" outlined color="blue-grey lighten-5">
+                <v-card class="mt-6" outlined color="grey lighten-3">
                   <p class="subtitle-1 font-weight-bold text-center">
                     Preprocessing
                   </p>
@@ -170,7 +170,7 @@
                 Cell selection
               </v-expansion-panel-header>
               <v-expansion-panel-content>
-                <v-card class="mt-6" outlined color="blue-grey lighten-5">
+                <v-card class="mt-6" outlined color="grey lighten-3">
                   <p class="subtitle-1 font-weight-bold text-center">
                     Cell selection
                   </p>